gpu-service/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "atlas-vox-gpu-service"
version = "0.1.0"
description = "GPU-accelerated TTS provider service for Atlas Vox"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "structlog>=24.0.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.9",
    "soundfile>=0.12.0",
    "numpy>=2.0.0",
    "torch>=2.1.0",
    "torchaudio>=2.1.0",
]

[project.optional-dependencies]
fish-speech = ["fish-speech>=0.1.0"]
chatterbox = ["chatterbox-tts"]
f5-tts = ["f5-tts"]
openvoice = ["openvoice"]
all = [
    "atlas-vox-gpu-service[fish-speech,chatterbox,f5-tts,openvoice]",
]

[tool.ruff]
target-version = "py311"
line-length = 120
